Controllers views

+++ src/main/java/com/example/springjobboard/controller/VacancyController.java
@@ -1,10 +1,13 @@
 package com.example.springjobboard.controller;
 
+import com.example.springjobboard.controller.util.ControllersUtil;
 import com.example.springjobboard.model.jobs.ResponseToVacancy;
 import com.example.springjobboard.model.jobs.Vacancy;
+import com.example.springjobboard.model.users.Applicant;
+import com.example.springjobboard.model.users.Employer;
 import com.example.springjobboard.repository.EmployerRepository;
 import com.example.springjobboard.repository.VacancyRepository;
-import com.example.springjobboard.utils.UsersUtil;
+import com.example.springjobboard.util.UsersUtil;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -27,14 +30,32 @@ public class VacancyController {
     @GetMapping
     public String getAll(Model model) {
         model.addAttribute("vacancies", vacancyRepository.findAll());
+        model.addAttribute("currentEmployer", getCurrentEmployer());
         return "vacancies/vacancies";
     }
 
     @GetMapping("/employer/{employerId}")
     public String getAllByEmployer(@PathVariable Long employerId, Model model) {
+
         var employer = employerRepository.findByIdEagerly(employerId, "vacancies");
+
         model.addAttribute("employer", employer);
         model.addAttribute("vacancies", employer.getVacancies());
+
+        return "vacancies/vacancies";
+    }
+
+    @GetMapping("/for-current-applicant")
+    public String getAllForCurrentApplicant(Model model) {
+
+        var currentApplicant = getCurrentApplicant();
+        if (currentApplicant == null) {
+            return "vacancies/vacancies";
+        }
+
+        model.addAttribute("filter", "suit your skills");
+        model.addAttribute("vacancies", vacancyRepository.getForApplicantBySkills(currentApplicant));
+
         return "vacancies/vacancies";
     }
 
@@ -44,14 +65,25 @@ public class VacancyController {
         var vacancy = getVacancyByIdOrThrowException(id, true);
         var optionalProperties = controllersUtil.getAllOptionalPropertiesWithValuesForEntity(vacancy);
 
+        var currentApplicantResponse = vacancy.getResponses().stream()
+                .filter(r -> r.getApplicant().equals(getCurrentApplicant()))
+                .findAny()
+                .orElse(null);
+
         model.addAttribute("vacancy", vacancy);
         model.addAttribute("optionalProperties", optionalProperties);
+        model.addAttribute("postedByCurrentEmployer", vacancy.getEmployer().equals(getCurrentEmployer()));
+        model.addAttribute("currentApplicantResponse", currentApplicantResponse);
 
         return "vacancies/vacancy";
     }
 
     @GetMapping("/new")
     public String initCreation(Model model) {
+
+        if (getCurrentEmployer() == null) {
+            return "redirect:/employers/new";
+        }
 
         var vacancy = new Vacancy();
         var optionalProperties = controllersUtil.getAllOptionalPropertiesForEntity(vacancy);
@@ -65,6 +97,10 @@ public class VacancyController {
     @PostMapping
     public String processCreation(Model model,
                                   @ModelAttribute @Valid Vacancy vacancy, BindingResult bindingResult) {
+
+        if (getCurrentEmployer() == null) {
+            return "redirect:/employers/new";
+        }
 
         var currentEmployer = usersUtil.getCurrentEmployer();
 
@@ -89,8 +125,16 @@ public class VacancyController {
     @GetMapping("/{id}/edit")
     public String initUpdate(@PathVariable Long id, Model model) {
 
+        if (getCurrentEmployer() == null) {
+            return "redirect:/employers/new";
+        }
+
         var vacancy = getVacancyByIdOrThrowException(id, true);
         var optionalProperties = controllersUtil.getAllOptionalPropertiesWithValuesForEntity(vacancy);
+
+        if (!vacancy.getEmployer().equals(getCurrentEmployer())) {
+            throw new RuntimeException("You have no permissions to edit other's vacancy");
+        }
 
         model.addAttribute("vacancy", vacancy);
         model.addAttribute("optionalProperties", optionalProperties);
@@ -102,7 +146,17 @@ public class VacancyController {
     public String processUpdate(@PathVariable Long id, Model model,
                                 @ModelAttribute @Valid Vacancy vacancy, BindingResult bindingResult) {
 
+        var currentEmployer = getCurrentEmployer();
+
+        if (currentEmployer == null) {
+            return "redirect:/employers/new";
+        }
+
         var vacancyPersisted = getVacancyByIdOrThrowException(id, true);
+
+        if (!vacancyPersisted.getEmployer().equals(currentEmployer)) {
+            throw new RuntimeException("You have no permissions to edit other's vacancy");
+        }
 
         if (bindingResult.hasErrors()) {
             model.addAttribute("errors", controllersUtil.bindingResultErrorsToMap(bindingResult));
@@ -125,6 +179,11 @@ public class VacancyController {
 
     @PostMapping("/{id}/delete")
     public String processDelete(@PathVariable Long id) {
+
+        if (getCurrentEmployer().getVacancies().stream().noneMatch(v -> v.getId().equals(id))) {
+            throw new RuntimeException("You have no permissions to edit other's vacancy");
+        }
+
         var result = vacancyRepository.deleteById(id);
         if (!result) {
             throw new EntityNotFoundException(String.format("vacancy with id '%d' is not found", id));
@@ -137,10 +196,13 @@ public class VacancyController {
 
         var vacancy = getVacancyByIdOrThrowException(id, true);
 
-        var currentApplicant = usersUtil.getCurrentApplicant();
-
+        var currentApplicant = getCurrentApplicant();
         if (currentApplicant == null) {
             throw new RuntimeException("You are not authorized as applicant");
+        }
+
+        if (vacancy.getEmployer().equals(getCurrentEmployer())) {
+            throw new RuntimeException("You can not respond to your own vacancy");
         }
 
         var response = new ResponseToVacancy();
@@ -169,5 +231,15 @@ public class VacancyController {
             throw new EntityNotFoundException(String.format("vacancy with id '%d' is not found", id));
         }
         return vacancy;
+    }
+
+    @ModelAttribute("currentEmployer")
+    private Employer getCurrentEmployer() {
+        return usersUtil.getCurrentEmployer();
+    }
+
+    @ModelAttribute("currentApplicant")
+    private Applicant getCurrentApplicant() {
+        return usersUtil.getCurrentApplicant();
     }
 }
